migrated service method to BlockedUser, and fixed tests
