+++ src/main/java/com/danielagapov/spawn/Services/User/IUserService.java
@@ -95,6 +95,4 @@ public interface IUserService {
     BaseUserDTO updateUser(UUID id, String bio, String username, String firstName, String lastName);
 
     SearchedUserResult getRecommendedFriendsBySearch(UUID requestingUserId, String searchQuery);
-    void removeFriendshipBetweenUsers(UUID userAId, UUID userBId);
-
 }
