+++ src/test/java/com/danielagapov/spawn/ServiceTests/BlockedUserServiceTests.java
@@ -9,6 +9,7 @@ import com.danielagapov.spawn.Models.User;
 import com.danielagapov.spawn.Repositories.IBlockedUserRepository;
 import com.danielagapov.spawn.Services.BlockedUser.BlockedUserService;
 import com.danielagapov.spawn.Services.FriendRequest.IFriendRequestService;
+import com.danielagapov.spawn.Services.FriendTag.IFriendTagService;
 import com.danielagapov.spawn.Services.User.IUserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -25,7 +26,11 @@ public class BlockedUserServiceTests {
     @Mock private IBlockedUserRepository blockedRepo;
     @Mock private IUserService userService;
     @Mock private IFriendRequestService friendRequestService;
+
     @Mock private ILogger logger;
+
+    @Mock
+    private IFriendTagService friendTagService;
 
     @InjectMocks private BlockedUserService blockedUserService;
 
@@ -63,10 +68,10 @@ public class BlockedUserServiceTests {
         when(blockedRepo.existsByBlocker_IdAndBlocked_Id(blockerId, blockedId)).thenReturn(false);
         when(userService.getUserEntityById(blockerId)).thenReturn(blocker);
         when(userService.getUserEntityById(blockedId)).thenReturn(blocked);
+        when(friendTagService.getFriendTagsByOwnerId(any())).thenReturn(List.of()); // ✅ add this stub
 
         blockedUserService.blockUser(blockerId, blockedId, "Testing");
 
-        verify(userService).removeFriendshipBetweenUsers(blockerId, blockedId);
         verify(blockedRepo).save(any(BlockedUser.class));
     }
 
@@ -75,6 +80,7 @@ public class BlockedUserServiceTests {
         when(blockedRepo.existsByBlocker_IdAndBlocked_Id(blockerId, blockedId)).thenReturn(false);
         when(userService.getUserEntityById(blockerId)).thenReturn(blocker);
         when(userService.getUserEntityById(blockedId)).thenReturn(blocked);
+        when(friendTagService.getFriendTagsByOwnerId(any())).thenReturn(List.of());
         doThrow(new DataAccessException("DB error") {}).when(blockedRepo).save(any());
 
         assertThrows(BaseSaveException.class, () ->
